dclock: add ability to monitor the battery state and flash the clock
We're now able to monitor battery status for stupid laptops like the
Lenovo t430s on which the battery led is placed *behind* the laptop
and you only know it was low when it suddenly cuts.

By specifying the path to a battery in /sys, each second the state is
checked and if less than 10 minutes remain, the digits flash red twice
a second in order to catch attention.

Example:
  dclock -bp /sys/class/power_supply/BAT0